Ex: Fixed error in X and corrected it by Y
Ex: The K-S test to compare two distributions is commonly used by coming up with a normal distribution that matches the mean, etc. of the data and comparing this theoretical dist. with your actual data. However, this is incorrect -- the data should be compared to a distribution that is determined intrinsically, or, if using the data, bootstrapping can be used. This paper or tutorial is helpful (link). I correct the error to use bootstrapping.