Created bootstrap subsampling
While scipy has a bootstrap function, it doesn't enable subsampling and instead samples n datapoints from the original distribution, but randomly and with replacement. For a large dataset this crashes my kernel and contributes to oversampling. To resolve the issue, I am experimenting with smaller boots. It is not obvious how we should handle determining boot size -- I am thinking about a parameter sweep as the next step.