Completed first draft.
In this draft, I have elaborated on why 1) the Cramer von Mises test seems superior to KS or AD for our usual MD data 2) compared datasets x and y using a permutation bootstrap and 3) creased z as a union of x and y the results from 2 to random subsets of z. Finally, I have also incorporated cells to investigate how random subsampling of the data fixes issues of autocorrelation. This code is flexible, allowing the user to alter the size of the boots/samples. think that we should probably do a small parameter sweep in order to reveal when our conclusions are strong or break down. The size of the dataset where breakdown occurs I think is interesting, and I would like to formalize thoughts in this area. I think it says something about the importance of conclusions and sparse sampling of some states.