Corrected logical flaw in the autocorrelation of bootstraps
Initially, I plotted autocorrelation of the bootstrap samples without time ordering (and it was about zero). However, when I bootstrap on the index, reorder the index numerically, and sample the x and y distributions using those values and take the autocorrelations for those signals then they are non-trivial again. This is quite interesting. Also, it seems like the pattern flips -- the autocorrelation is actually higher for P210S bootstraps than ordered chunks of the signal (same chunk size). For WT the result is intuitive: autocorrelation is higher when n is sampled consecutively than when n is bootstrapped out. I would like to come up with a numerical way to evaluate these heuristics.